+++ doc/source/changes.rst
@@ -15,6 +15,8 @@ Norman-0.6.2
 
 -   Hooks added to `Table` to allow more control over validation.
 -   Add `Query.field`, allowing queries to traverse tables.
+-   Add `Query.add`, allowing records to be created based on query criteria
+
 
 Norman-0.6.1
 ------------